This is a major change to quite many locations in the code. As a result, I hope I have achieved the following steps:
- The new 'gui' tool replaces the old cltool completely

- The capacities and features of the synth are defined per-build

- Working stuff from the old tools/ directory has been included in
  the new gui/ directory (mostly just copy-paste, unfortunately)

- Started a new example code targeting the Instanssi 2014 event.

Old toolchain is not working anymore, and the old examples don't
work either. As of yet, there's no converter between old and new
formats of patch and bank data.

The songdata export is not yet working after the revamps, but
after that gets fixed, it will be a proper time to remove dead
code and wrap up the whole synti2 project (and move on to synti3:))